% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_reg_off}
\alias{add_reg_off}
\title{Add a column with names of offshore regions}
\usage{
add_reg_off(x, regN, regN_off = paste0(regN, "_off"), offshore = "offshore")
}
\arguments{
\item{x}{data frame with required two columns: \code{offshore} and \code{regN}}

\item{regN}{character name of the column with region names}

\item{regN_off}{character name of the column with offshore region names,
that will be added to the data frame, default is \verb{\{regN\}_off}}

\item{offshore}{character name of the column with offshore indicator,
default is 'offshore'}
}
\value{
data frame with added column \code{regN_off}
}
\description{
Add a column with names of offshore regions
}
