% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_fugires.R
\name{add_tech_type}
\alias{add_tech_type}
\title{Add 'tech_type' column to a data.frame}
\usage{
add_tech_type(x, factors = TRUE, force = TRUE)
}
\arguments{
\item{x}{data.frame that has 'process' column}

\item{factors}{logical, if TRUE, the 'tech_type' column will have class}

\item{force}{logical, if TRUE will add unknown types of technologies to the factor level, or report the 'unknown types' error if FALSE.}
}
\value{
x data.frame with added or overwritten 'tech_type' column.
}
\description{
Add 'tech_type' column to a data.frame
}
