% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_vintage}
\alias{drop_vintage}
\title{Drop the numeric part of a string (aka vintage in names of technologies)}
\usage{
drop_vintage(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
The input character vector with a trailing \verb{_<number>} (vintage) suffix removed.
}
\description{
Drop the numeric part of a string (aka vintage in names of technologies)
}
\examples{
names(ideea_techs$ECOASUB)
drop_vintage(names(ideea_techs$ECOASUB))

}
