% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_extra.R
\name{get_ideea_cf}
\alias{get_ideea_cf}
\title{Solar and wind capacity factors by year, region and cluster}
\usage{
get_ideea_cf(
  resource = "sol",
  nreg = 5,
  year = 2019,
  tol = 0.05,
  data = "merra2",
  data_file = "merra_raw_{year}.fst",
  clusters_file = "locid_{resource}_cl_r{nreg}.RData",
  tsl_format = "d365_h24",
  load_from_extra = TRUE,
  save_to_extra = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{resource}{character, name of the resource: "sol" or "win"}

\item{nreg}{integer, the number of regions}

\item{year}{integer, the year of the data to return}

\item{tol}{numeric, the tolerance of the information loss in the clustering from 0 to 1. If 0 the number of clusters is equal to the number of locations in MERRA2 database. When tol is 1, there is only one cluster per region.}

\item{data}{character, the name of the data source, the directory in the \code{ideea_extra} directory. Currently only "merra2" is supported.}

\item{data_file}{character, the name of the file with the raw data in the \code{data} sub-directory in \code{ideea_extra}. Default is "merra_raw_{year}.fst", valid for MERRA2 data.}

\item{clusters_file}{character, the name of the file with the clusters in the \code{data} directory. Default is "locid_{resource}_cl_r{nreg}.RData", available in \code{ideea_data}.}

\item{tsl_format}{character, the format of the time-slices. Default is "d365_h24". If changed, the clustering will be done based on the new time-slices.}

\item{load_from_extra}{logical, if TRUE, the function will try to read the capacity factors from the file in the \code{data} sub-directory in \code{ideea_extra}. If the file with requested parameters exists, it will be read. If the file doesn't exist, the function will calculate the capacity factors. Default is TRUE.}

\item{save_to_extra}{logical, if TRUE, the function will save the estimated capacity factors to the file in the \code{data} sub-directory in \code{ideea_extra}. Default is TRUE.}

\item{overwrite}{logical, if TRUE, the function will overwrite the existing file with the capacity factors. Default is FALSE.}
}
\value{
data.table, the capacity factors for the requested year by region, cluster, and time-slice. The data.table contains the following columns:
\itemize{
\item \code{locid} - the location ID from the MERRA2 database,
\item \code{year} - the year of the data,
\item \code{slice} - the time-slice,
\item \code{cluster} - the cluster ID,
\item \verb{reg\{nreg\}} and \verb{reg\{nreg\}_off} - the region ID and the offshore region ID,
\item \verb{wcf_*} or \verb{scf_*} - the capacity factor for the wind or solar resource at the given height or tracking type.
}
}
\description{
Solar and wind capacity factors by year, region and cluster
}
\examples{
# don't run
# default tolerance level (0.05)
get_ideea_cf(resource = "sol", nreg = 5, year = 2019)
get_ideea_cf(resource = "sol", nreg = 32, year = 2019)

get_ideea_cf(resource = "win", nreg = 5, year = 2019)
get_ideea_cf(resource = "win", nreg = 32, year = 2019)

# custom tolerance level (0.01)
get_ideea_cf(resource = "sol", nreg = 5, tol = 0.01, year = 2019)
get_ideea_cf(resource = "sol", nreg = 32, tol = 0.01, year = 2019)

get_ideea_cf(resource = "win", nreg = 5, tol = 0.01, year = 2019)
get_ideea_cf(resource = "win", nreg = 32, tol = 0.01, year = 2019)
}
