% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_extra.R
\name{get_ideea_cl_sf}
\alias{get_ideea_cl_sf}
\title{Load sf (simple feature) object of solar or wind clusters}
\usage{
get_ideea_cl_sf(
  resource = "sol",
  nreg = 5,
  tol = 0.05,
  data = "merra2",
  clusters_sf_file = "locid_{resource}_cl_r{nreg}_sf.RData"
)
}
\arguments{
\item{resource}{character, name of the resource: "sol" or "win"}

\item{nreg}{integer, the number of regions}

\item{tol}{numeric, the tolerance of the information loss in the clustering from 0 to 1. If 0 the number of clusters is equal to the number of locations in MERRA2 database. When tol is 1, there is only one cluster per region.}

\item{data}{character, the name of the data source, the directory in the \code{ideea_extra} directory. Currently only "merra2" is supported.}

\item{clusters_sf_file}{character, the name of the file with the clusters in the \code{data} directory. Default is "locid_{resource}_cl_r{nreg}_sf.RData", available in \code{ideea_data}.}
}
\value{
sf, the simple feature object with the clusters for the requested resource.
}
\description{
Load sf (simple feature) object of solar or wind clusters
}
\examples{
# don't run
get_ideea_cl_sf(resource = "sol", nreg = 5)
get_ideea_cl_sf(resource = "win", nreg = 32, tol = 0.01)
}
