% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ideea_map}
\alias{get_ideea_map}
\title{Get IDEEA map with custom number of regions}
\usage{
get_ideea_map(
  nreg = 5,
  offshore = FALSE,
  islands = TRUE,
  aggregate = (nreg != 46),
  rename = FALSE,
  reg_off = offshore,
  merge_islands = TRUE,
  ...
)
}
\arguments{
\item{nreg}{integer, number of regions, one of 1, 5, 7, 32, 34, 36, 37 or 46}

\item{offshore}{logical, if TRUE, offshore area with associated regions will be returned}

\item{islands}{logical, should "Andaman and Nicobar" ("AN") and "Lakshadweep" ("LD") be included.}

\item{aggregate}{logical, if true, geometries will be aggregated by region.}

\item{rename}{logical, if TRUE, the \code{region} and \code{name} will be returned instead of \verb{reg\{nreg\}} and \verb{name\{nreg\}}}

\item{reg_off}{logical, if TRUE, \code{reg_off} or \verb{reg\{nreg\}_off} column will be added with added \verb{_off} to the names of offshore areas/regions.}

\item{merge_islands}{logical, \code{mainland} column should be dropped (if TRUE) or added (FALSE) to the returned sf-object.}

\item{...}{additional arguments (currently unused).}
}
\value{
IDEEA map with in 'sf' or 'sp' format
}
\description{
Get IDEEA map with custom number of regions
}
\examples{
library(sf)
gis_sf <- get_ideea_map(nreg = 1, offshore = F, islands = T)
plot(gis_sf[1], key.width = lcm(4))
gis_sf <- get_ideea_map(nreg = 7, offshore = T, islands = T, reg_off = T)
plot(gis_sf["region"], key.width = lcm(4))
plot(gis_sf["reg_off"], key.width = lcm(4))
gis_sf <- get_ideea_map(nreg = 32, offshore = T, islands = T)
plot(gis_sf[1], key.width = lcm(4))
}
