% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideea_data}
\alias{ideea_data}
\title{IDEEA datasets}
\format{
list of data frames:
\describe{
\item{reg_tbl}{mapping table of regions}
\item{load_2019_MWh}{load curve by region in 2019}
\item{elc_consumption}{Historic and projected electricity consumption by region}
\item{coal}{coal domestic resources by region}
\item{lignite}{lignite domestic resources by region}
\item{oil}{oil domestic resources}
\item{gas}{natural gas domestic resources}
\item{biomass}{biomass domestic resources}
\item{generators}{installed power plants capacity by fuel type and region}
\item{merra_raw_2014}{MERRA2 subset for 2014}
...
}
}
\usage{
data(ideea_data)
}
\description{
Embedded datasets for IDEEA models
}
\keyword{datasets}
