% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_gif.R
\name{ideea_gif_cf}
\alias{ideea_gif_cf}
\title{Create GIF animation of capacity factors by cluster and time-slice}
\usage{
ideea_gif_cf(
  x,
  ideea_cl_sf,
  ideea_sf = NULL,
  cf_name = names(x)[grepl("cf_"), names(x)][1],
  slice = unique(x$slice)[1:24],
  timestamp.stamp = TRUE,
  fill_scale_lims = range(x[[cf_name]]),
  fps = 12,
  gif.width = 576,
  gif.height = 576,
  filename = "ideea_cl.gif"
)
}
\arguments{
\item{x}{data.frame of capacity factors, typically from get_ideea_cf()}

\item{ideea_cl_sf}{sf, cluster shapefile, typically from get_ideea_cl_sf()}

\item{ideea_sf}{sf, ideea shapefile, typically from get_ideea_map()}

\item{cf_name}{character, name of capacity factor column in x (e.g. "wcf_100m", "scf_tl", etc.)}

\item{slice}{character vector, time-slices to udpate in the GIF}

\item{timestamp.stamp}{logical, if TRUE, add timestamp to the plot}

\item{fill_scale_lims}{numeric vector of length two, limits of the fill scale}

\item{fps}{numeric, frames per second, default is 12}

\item{gif.width}{numeric, width of the GIF, default is 576}

\item{gif.height}{numeric, height of the GIF, default is 576}

\item{filename}{character, name of the GIF file}
}
\value{
A GIF animation of capacity factors by cluster and time-slice saved in the working directory
}
\description{
Create GIF animation of capacity factors by cluster and time-slice
}
\examples{
# do not run
resource <- "win"; cf_name <- "wcf_100m"
resource <- "sol"; cf_name <- "scf_tl"
nreg <- 5
tol <- 0.01

ideea_sf <- get_ideea_map(nreg = nreg, offshore = T, islands = T)
ideea_cl_sf <- get_ideea_cl_sf(resource = resource, tol = tol)

ideea_cl_sf$cluster |> unique()
plot(ideea_cl_sf["cluster"])

x <- get_ideea_cf(resource, tol = tol)

slices_1day_per_month <-
  ideea_modules$electricity$reg7_base$partial_calendar_1day_per_month@timetable$slice

ideea_gif_cf(x, ideea_cl_sf, ideea_sf, cf_name = cf_name,
             slice = slices_1day_per_month,
             filename = glue("tmp/{resource}.gif"))
}
