% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ideea_global_options}
\alias{ideea_global_options}
\title{Save, load, edit, or remove IDEEA global options}
\usage{
ideea_global_options(edit = FALSE, load = !edit, remove = FALSE)
}
\arguments{
\item{edit}{logical, if TRUE open the file in the IDE or default text editor}

\item{load}{logical, if TRUE load the global options from the file}

\item{remove}{logical, if TRUE remove the global options file}
}
\value{
on the first run creates file in '~/.ideea.R' with commented settings and opens the file for editing. On the second run, sources the file to load the settings. Returns TRUE invisibly if successful, or FALSE if not. The function is called with the loading of IDEEA package and sources the '~/.ideea.R' file if previously created.
}
\description{
Save, load, edit, or remove IDEEA global options
}
