% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideea_modules}
\alias{ideea_modules}
\title{IDEEA model components}
\format{
list of data frames:
\describe{
\item{energy}{Primary energy supply module (see article \code{energy})}
\item{electricity}{Electric power sector module}
...
}
}
\usage{
ideea_modules
}
\description{
Model modules and components
}
\keyword{datasets}
