% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ideea_options_summary}
\alias{ideea_options_summary}
\title{Summary of IDEEA global options}
\usage{
ideea_options_summary()
}
\value{
a list of IDEEA global options
}
\description{
Summary of IDEEA global options
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ideea_options_summary()
\dontshow{\}) # examplesIf}
}
