% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_scenarios.R
\name{ideea_scenario_dir_name}
\alias{ideea_scenario_dir_name}
\title{Add model, calendar, and other details to a name IDEEA scenario}
\usage{
ideea_scenario_dir_name(
  name,
  model_name = NULL,
  calendar_name = NULL,
  horizon_name = NULL,
  suffix = NULL,
  sep = "_"
)
}
\arguments{
\item{name}{character, basic name of IDEEA scenario}

\item{model_name}{character, model name}

\item{calendar_name}{character, calendar name}

\item{suffix}{character, additional string to add to the name}

\item{sep}{character, a separator between identifiers, "_" by default}
}
\description{
Add model, calendar, and other details to a name IDEEA scenario
}
