% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_scenarios.R
\name{ideea_scenarios}
\alias{ideea_scenarios}
\title{Get current IDEEA-scenarios directory}
\usage{
ideea_scenarios()
}
\value{
Character path to the current IDEEA scenarios directory.
}
\description{
Get current IDEEA-scenarios directory
}
