% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_scenarios.R
\name{ideea_scenarios_list}
\alias{ideea_scenarios_list}
\title{List saved scenarios in the IDEEA-scenarios directory}
\usage{
ideea_scenarios_list()
}
\value{
A character vector of saved scenario directory names.
}
\description{
List saved scenarios in the IDEEA-scenarios directory
}
