% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_snapshot.R
\name{ideea_snapshot_cf}
\alias{ideea_snapshot_cf}
\title{Plot a snapshot of capacity factors by cluster and time-slice}
\usage{
ideea_snapshot_cf(
  x,
  ideea_cl_sf,
  ideea_sf = NULL,
  cf_name = names(x)[grepl("cf_"), names(x)][1],
  slice = sample(x$slice, 1),
  timestamp.stamp = TRUE,
  return_data = FALSE,
  fill_scale_lims = range(x[[cf_name]]),
  ...
)
}
\arguments{
\item{x}{data.frame, capacity factors data, an outpuf of \code{get_ideea_cf()}}

\item{ideea_cl_sf}{sf object, cluster shapefile, an output of \code{get_ideea_cl_sf()}}

\item{ideea_sf}{sf object, ideea shapefile, an output of \code{get_ideea_map()}}

\item{cf_name}{character, the name of the capacity factor column (e.g. "wcf_100m", "scf_tl", etc.)}

\item{slice}{character of length one, the time-slice to plot (e.g. "d001_h00")}

\item{return_data}{logical, if TRUE, the function returns the data used for plotting}

\item{fill_scale_lims}{numeric of length two, the limits of the fill scale}

\item{...}{ignored}

\item{datatime}{logical, if TRUE, the time-slice is converted to a timestamp}
}
\value{
A ggplot object of capacity factors by cluster and time-slice or the data used for plotting
}
\description{
Plot a snapshot of capacity factors by cluster and time-slice
}
\examples{
resource <- "win"
nreg <- 5
tol <- 0.1
cf_name <- "wcf_100m"
ideea_sf <- get_ideea_map(nreg = nreg, offshore = T, islands = T)
ideea_cl_sf <- get_ideea_cl_sf(resource = resource, tol = tol)
x <- get_ideea_cf(resource, tol = tol)
ideea_snapshot_cf(x, ideea_cl_sf, ideea_sf, cf_name = cf_name)
}
