% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_extra.R
\name{set_ideea_extra}
\alias{set_ideea_extra}
\title{Specify directory for IDEEA external dataset (\code{ideea_extra})}
\usage{
set_ideea_extra(path = NULL)
}
\arguments{
\item{path}{character, path to the dataset}
}
\value{
(invisible) logical, TRUE if the directory is set, FALSE otherwise.
}
\description{
Specify directory for IDEEA external dataset (\code{ideea_extra})
}
