% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideea_scenarios.R
\name{set_ideea_scenarios}
\alias{set_ideea_scenarios}
\title{Set directory for IDEEA scenarios}
\usage{
set_ideea_scenarios(path = NULL)
}
\arguments{
\item{path}{character path to the directory with/for IDEEA scenarios}
}
\description{
Set directory for IDEEA scenarios
}
