% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity.R
\name{drop_cluster}
\alias{drop_cluster}
\title{Drop cluster identifier from a name}
\usage{
drop_cluster(name, digits = 2, cluster_flag = "CL")
}
\arguments{
\item{cluster_flag}{}
}
\description{
Drop cluster identifier from a name
}
