% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity.R
\name{drop_process_cluster}
\alias{drop_process_cluster}
\title{dat |> drop_process_vintage()}
\usage{
drop_process_cluster(x, digits = 2, cluster_flag = "CL")
}
\description{
dat |> drop_process_vintage()
}
