% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_process_vintage}
\alias{drop_process_vintage}
\title{Drop the numeric part of a names in \code{process} column of a data frame}
\usage{
drop_process_vintage(x)
}
\arguments{
\item{x}{}
}
\description{
Drop the numeric part of a names in \code{process} column of a data frame
}
\examples{
dat <- data.frame(scenario = "BASE", process = names(ideea_techs$ENUC))
dat
dat |> drop_process_vintage()
}
