% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cluster}
\alias{get_cluster}
\title{Extract cluster string or number from a name}
\usage{
get_cluster(name, digits = 2, cluster_flag = "CL")
}
\arguments{
\item{cluster_flag}{}
}
\description{
Extract cluster string or number from a name
}
