% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ideea_data}
\alias{get_ideea_data}
\title{Get data from IDEEA database}
\usage{
get_ideea_data(
  name,
  nreg = 7,
  variable,
  agg_fun = sum,
  raw = FALSE,
  offshore = FALSE,
  islands = FALSE
)
}
\arguments{
\item{name}{character, name of data table, one of: "coal", "oil", "gas", "biomass", "lignite", ...}

\item{nreg}{integer, number of region to return}

\item{variable}{name or a part of name (regular expression) of variable to return.}

\item{agg_fun}{character, function to aggregate data by region, default "sum" (for volumes), "mean" is advised for costs variable.}

\item{raw}{logical, should the raw table be returned, FALSE by default}

\item{offshore}{logical, should the data for offshore areas be returned, FALSE by default}

\item{islands}{logical, should the data for remote islands be returned, FALSE by default}
}
\description{
Get data from IDEEA database
}
\examples{
  get_ideea_data("coal", raw = T)
  get_ideea_data("coal", nreg = 7, "reserve")
  get_ideea_data("oil", nreg = 34, "reserve", islands = T)
  get_ideea_data("coal", nreg = 7, "cost", agg_fun = mean)
  get_ideea_data("merra_raw_2014", raw = T) \%>\% head()
}
