% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEEA-package.R
\docType{package}
\name{IDEEA-package}
\alias{IDEEA}
\alias{IDEEA-package}
\title{IDEEA: Indian Zero Carbon Energy Pathways (IDEEA) optimization model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

IDEEA is a capacity expansion and energy resource planning model built with a goal to explore energy transition options and pathways for India, and inform policy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ideea-model}
  \item \url{https://ideea-model.github.io/IDEEA/index.html}
  \item Report bugs at \url{https://github.com/ideea-model/IDEEA/issues}
}

}
\author{
\strong{Maintainer}: IDEEA comunity \email{ideea@email.org}

}
\keyword{internal}
