% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_with_cluster}
\alias{name_with_cluster}
\title{Combine name with cluster number}
\usage{
name_with_cluster(name, cluster, ndigits = 2)
}
\arguments{
\item{ndigits}{}
}
\description{
Combine name with cluster number
}
